InfoMix/requirements.txt
cusignal==22.10.1
ffmpeg==1.4
ffmpeg_python==0.2.0
HyperPyYAML==1.2.0
librosa==0.9.1
mir_eval==0.7
numpy==1.21.0
pandas==1.4.2
praat_textgrids==1.4.0
pydub==0.25.1
pypesq==1.2.4
scipy==1.9.1
soundfile==0.12.1
SoundFile==0.10.3.post1
speechbrain==0.5.14
torch==1.11.0
torchaudio==0.11.0
tqdm==4.64.0
